Updates controllers and mailers
