+++ app/controllers/api/v1/consumers_controller.rb
@@ -1,0 +1,22 @@
+module Api
+  module V1
+    class ConsumersController < ApiController
+      CONSUMER_ROOT = 'consumer'.freeze
+      load_and_authorize_resource
+
+      def update
+        @consumer.update!(update_params)
+
+        render json: ConsumerSerializer.render(@consumer, root: CONSUMER_ROOT), 
+              status: :ok
+      end
+
+      private
+
+      def update_params
+        params.require(:consumer).permit(:first_name,
+                                         :last_name)
+      end
+    end
+  end
+end
