--- app/controllers/user_token_controller.rb
@@ -1,2 +1,0 @@
-class UserTokenController < Knock::AuthTokenController
-end
