+++ app/mailers/consumer_mailer.rb
@@ -5,10 +5,11 @@ class ConsumerMailer < BaseMandrillMailer
   def password_reset(user_id, jwt)
     user = User.find(user_id)
     token = jwt
-    role = user.profile_type
+    profile = user.profile
     email = user.email
     merge_vars = {
-      'RESET_LINK' => "#{ENV['EMAIL_ORIGIN']}/password-reset?token=#{token}&role=#{role}"
+      'RESET_LINK' => "#{ENV["EMAIL_ORIGIN"]}/passwords/#{user.id}/edit?token=#{token}",
+      'USER_NAME' => profile.first_name.capitalize
     }
 
     body = mandrill_template('password-reset', merge_vars)
